Some reordering
